// interfaces/resultUser-interface.ts
// Generated by https://quicktype.io

export interface ResultUserInterface {
    ListTarjets: ListTarjet[];
}

export interface ListTarjet {
    IdUsuario:      string;
    NombreCompleto: string;
    Cargo:          string;
    Actividad:      string;
    FondoF:         string;
    Activa:         number;
    SegmentoId:     string;
    SegmentoDesc:   string;
    Titulo:         string;
    TituloDesc:     string;
    Token:          string;
    ImgFoto:        string;
    Tipo:           string;
    PublicPriva:    number;
    Alias:          string;
    RegistroTarjet: boolean;
    Coordenadas:    string;
    Distancia:      string;
    RangoLocal:     string;
}